Set global git identity for GH actions test env
GitHub actions test env errors with `Error: Author identity unknown`
for rebase tests if user.name & user.email are not set. This commit
sets a global identity for tests to satisfy GH actions. As the tests
are not run in a Docker env, the original global identity, if any,
must be restored following rebase tests. This is not ideal, and a
better solution to avoid accidentally messing with the dev env would be
to run the tests in a container.

This improvement can be in a separate pull request along with the GPG signing test
case, as the bare minimum feature set for this pull request has been met.

Signed-off-by: Christina Ying Wang <christina@balena.io>