Skip --force-with-lease test for now
While the functionality where a failed --force-with-lease push throwing
a RemoteChangedError has been tested manually and confirmed to work, it has
been difficult to test locally due to the race condition between running the rebase
and separately pushing a change to remote. simple-git's rebase has also been
resistant to function mocking. More thought is required to get this test to work.

Signed-off-by: Christina Ying Wang <christina@balena.io>